X00TD: displayconfig: Add brightness ramp rate
Android 12 allows the brightness ramp rate to be changed separately
for fast/slow and increase/decrease in float. Once loaded,
DisplayDeviceConfig has higher priority than the integer config
values in frameworks. The format definition is in [1].

The value can be found in boot logs by grepping DisplayDeviceInfo.
On android 11, boot log reports the 19xxxxx value of the second
variant. On android 12, boot log reports the 46xxxxx value of the
first variant.

The xml file can be put in either product or vendor.

[1] frameworks/base/services/core/xsd/display-device-config/display-device-config.xsd

[clarencelol: Adjust accordingly based on brightness minimum and nits in overlay]
Signed-off-by: clarencelol <clarencekuiek@icloud.com>
Signed-off-by: karthik1896 <karthik1896@outlook.com>